Move parse_init_payload and Lcg from benchmark to the edgeless_function crate
